as - refactor ChatToggle and ChatContainer into their own react components
