+++ frontend/src/main/pages/PlayPage.js
@@ -1,5 +1,4 @@
-import React, {useState} from "react";
-import { Container, CardGroup, Button } from "react-bootstrap";
+import { Container, CardGroup } from "react-bootstrap";
 import { useParams } from "react-router-dom";
 
 import BasicLayout from "main/layouts/BasicLayout/BasicLayout";
@@ -11,7 +10,7 @@ import Profits from "main/components/Commons/Profits";
 import { useBackend, useBackendMutation } from "main/utils/useBackend";
 import { useCurrentUser } from "main/utils/currentUser";
 import Background from "../../assets/PlayPageBackground.png";
-import ChatPanel from "main/components/Chat/ChatPanel";
+import ChatContainer from "main/components/Chat/ChatContainer";
 
 export default function PlayPage() {
 
@@ -115,41 +114,6 @@ const onSuccessSell = () => {
     mutationsell.mutate(userCommons)
   };
 
-  const [isChatOpen, setIsChatOpen] = useState(false);
-
-  const toggleChatWindow = () => {
-    setIsChatOpen((prevState) => !prevState);
-  };
-
-  // const chatButtonStyle = {
-  //   width: '60px',
-  //   height: '60px',
-  //   borderRadius: '50%',
-  //   color: 'black',
-  //   position: 'fixed',
-  //   right: '20px',
-  //   fontSize: '24px',
-  // };
-
-  const chatButtonStyle = {
-    width: '60px',
-    height: '60px',
-    borderRadius: '50%',
-    color: 'black',
-    position: 'fixed',
-    bottom: '20px',
-    right: '20px',
-    fontSize: '24px',
-    boxShadow: '0px 4px 20px rgba(0, 0, 0, 0.75)',
-  };
-
-  const chatContainerStyle = {
-    width: '550px',
-    position: 'fixed',
-    bottom: '100px',
-    right: '20px',
-  };
-
   return (
     <div style={{ backgroundSize: 'cover', backgroundImage: `url(${Background})` }} data-testid="playpage-div">
       <BasicLayout >
@@ -166,12 +130,7 @@ const onSuccessSell = () => {
           }
         </Container>
       </BasicLayout>
-      <div style={chatContainerStyle} data-testid="playpage-chat-div">
-        {!!isChatOpen && <ChatPanel commonsId={commonsId}/>}
-        <Button style={chatButtonStyle} onClick={toggleChatWindow} data-testid="playpage-chat-toggle">
-            {!!isChatOpen ? '🔻' : '💬'}
-        </Button>
-      </div>
+      <ChatContainer commonsId={commonsId} />
     </div>
   )
 }
